requirements-feast.txt
feast[postgres,aws]==0.36.0
protobuf<4.23.4,>3.20
redis==4.2.0
hiredis==2.0.0
dask[dataframe]==2023.1.0
psycopg2-binary
fsspec==2023.12.2
s3fs==2023.12.2
pandas==1.4.4
pyarrow==12.0.1
numpy==1.24.4
scipy==1.12.0
pyspark>=3.3.0 
scikit-learn==1.5.1

"""pip freeze 했을때 
aiobotocore==2.23.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
boto3==1.39.8
botocore==1.39.8
bowler==0.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dask==2023.1.0
databricks-sdk==0.61.0
Deprecated==1.2.18
dill==0.3.9
docker==7.1.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.116.1
fastavro==1.12.0
feast==0.36.0
fire==0.7.0
fissix==24.4.24
Flask==3.1.1
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2023.12.2
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.3
grpcio==1.74.0
grpcio-health-checking==1.62.3
grpcio-reflection==1.62.3
grpcio-tools==1.62.3
gunicorn==23.0.0
h11==0.16.0
hiredis==2.0.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
icecream==2.1.5
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
locket==1.0.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.5
mlflow==3.1.4
mlflow-skinny==3.1.4
mmh3==5.2.0
moreorless==0.5.0
multidict==6.6.3
mypy==1.17.1
mypy-protobuf==3.1.0
mypy_extensions==1.1.0
numpy==1.24.4
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==25.0
pandas==1.4.4
pandavro==1.5.2
partd==1.4.2
pathspec==0.12.1
pillow==11.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.23.3
psycopg2-binary==2.9.10
py4j==0.10.9.9
pyarrow==12.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pyspark==4.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==4.2.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
rsa==4.9.1
s3fs==2023.12.2
s3transfer==0.13.1
scikit-learn==1.7.1
scipy==1.12.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==1.4.54
sqlalchemy2-stubs==0.0.2a38
sqlparse==0.5.3
starlette==0.47.2
tabulate==0.9.0
tenacity==8.5.0
termcolor==3.1.0
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
tqdm==4.67.1
typeguard==4.4.4
types-protobuf==6.30.2.20250703
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
volatile==2.1.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
"""